requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
awsebcli==3.20.10
bcrypt==4.2.0
blinker==1.8.2
botocore==1.31.85
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
Flask==2.3.2
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
multidict==6.0.5
pathspec==0.10.1
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.8.5
six==1.16.0
SQLAlchemy==2.0.30
termcolor==1.1.0
twilio==9.2.4
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.1.9
Werkzeug==3.0.3
yarl==1.9.4
python-dotenv==1.0.1
gunicorn==23.0.0